style: fix refressions and improve budget styles
