+++ apps/picsa-tools/budget-tool/src/app/components/_deprecated/cell-editor.html
@@ -78,8 +78,7 @@
             (valueChanged)="onEditorChange($event, 'outputs')"
           ></budget-cell-editor-input-values>
           <!-- <next-button (click)="stepper.next()"></next-button -->
-          ></mat-step
-        >
+        </mat-step>
       </mat-horizontal-stepper>
     </mat-step>
     <!-- Consumed -->
@@ -105,8 +104,7 @@
             (valueChanged)="onEditorChange($event, 'produceConsumed')"
           ></budget-cell-editor-produce-consumed>
           <!-- <next-button (click)="stepper.next()"></next-button -->
-          ></mat-step
-        >
+        </mat-step>
       </mat-horizontal-stepper>
       <!-- ------------------------------------------------------------------- -->
     </mat-step>
