+++ apps/picsa-tools/budget-tool/src/app/components/card/card-new/card-new-dialog.scss
@@ -1,0 +1,13 @@
+:host {
+  display: block;
+}
+.budget-card {
+  width: 110px;
+  margin-bottom: 1rem;
+}
+.card-title {
+  height: 2.5em;
+  padding: 5px;
+  background: #dcdcdc;
+  overflow: auto;
+}
