+++ apps/picsa-tools/budget-tool/src/app/components/cell-editor/family-labour/family-labour.html
@@ -13,10 +13,10 @@
     </div>
 
     <mat-form-field class="card-input-field quantity member-input">
+      <mat-label>{{'Days' | translate}}</mat-label>
       <input
         matInput
         class="input-editable"
-        [placeholder]="'Days' | translate"
         [value]="familyCard.values ? familyCard.values.quantity : null"
         type="number"
         (change)="setValue($event, i)"
