+++ apps/picsa-tools/budget-tool/src/app/components/cell-editor/family-labour/family-labour.scss
@@ -5,6 +5,7 @@
 }
 .member-card {
   display: flex;
+  flex-direction: row;
   padding: 1rem;
   align-items: center;
   position: relative;
@@ -18,7 +19,7 @@
 }
 .member-card-container {
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(160px, 1fr));
+  grid-template-columns: repeat(auto-fill, minmax(210px, 1fr));
   gap: 0.5rem;
 }
 .member-delete-button {
@@ -32,11 +33,10 @@
 .member-input {
   flex: 1;
   margin: 0 1rem;
-  overflow: hidden;
 }
 .member-add-container {
   margin-top: 2rem;
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(160px, 1fr));
+  grid-template-columns: repeat(auto-fill, minmax(210px, 1fr));
   gap: 0.5rem;
 }
