+++ apps/picsa-tools/budget-tool/src/app/components/cell-editor/input-values/input-values.scss
@@ -9,10 +9,10 @@
 }
 
 .card-input-field.quantity {
-  width: 65px;
+  width: 175px;
 }
 .card-input-field.cost {
-  width: 135px;
+  width: 175px;
 }
 .card-input.total {
   width: 200px;
