+++ apps/picsa-tools/budget-tool/src/app/components/cell/cell.html
@@ -36,7 +36,7 @@
         <div *ngFor="let card of cellData">
           <budget-card-image class="x-small" [card]="card"></budget-card-image>
           <span
-            class="badge family-badge"
+            [class]="'badge family-badge '+card.id"
             [matBadge]="card.values!.quantity"
           ></span>
         </div>
