+++ apps/picsa-tools/budget-tool/src/app/components/list-item/budget-list-item.html
@@ -1,4 +1,7 @@
-<mat-card appearance="outlined" style="margin-bottom: 10px; cursor: pointer">
+<mat-card
+  appearance="raised"
+  style="margin-bottom: 10px; padding: 16px; cursor: pointer"
+>
   <div style="display: flex">
     <budget-card-image
       class="small"
