+++ apps/picsa-tools/budget-tool/src/app/components/list-item/budget-list-item.scss
@@ -1,1 +1,4 @@
-:host {}
+mat-card-content {
+  padding: 0;
+  margin-bottom: 0.5rem;
+}
