+++ apps/picsa-tools/budget-tool/src/app/pages/create/budget-create.page.scss
@@ -8,12 +8,12 @@ textarea {
 }
 
 .variable-container {
-  min-width: 80px;
-  font-size: 30px;
-  height: 80px;
   display: flex;
   flex-direction: column;
   justify-content: center;
+  min-width: 80px;
+  font-size: 30px;
+  height: 80px;
   margin: 0 20px 20px 0;
 }
 
@@ -21,6 +21,13 @@ textarea {
   color: var(--ion-color-primary);
   outline: 2px solid var(--ion-color-primary);
   font-weight: bold;
+}
+.variable-value {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  height: 100%;
+  min-width: 100px;
 }
 
 /* TODO(mdc-migration): The following rule targets internal classes of select that may no longer apply for the MDC version. */
