+++ libs/theme/src/_overrides.scss
@@ -26,12 +26,25 @@ mat-card-title.mat-mdc-card-title {
 mat-card-content.mat-mdc-card-content {
   font-size: 14px;
 }
-
-/*******************************************************************
-      Inputs
-******************************************************************/
-// TODO - check
-.mat-input-element.input-editable {
+// Add padding to all dialogs
+.mat-mdc-dialog-surface {
+  padding: 24px;
+}
+// Remove padding from dialog action buttons
+div.mdc-dialog__actions {
+  padding-left: 0;
+  padding-right: 0;
+}
+// Increase size and set primary colour for text inputs
+input.mat-mdc-input-element.mdc-text-field__input {
   font-size: 30px;
-  color: var(--color-primary);
+  color: var(--color-primary) !important;
+}
+// Reduce placeholder font size (inherited above)
+.mdc-text-field__input::placeholder {
+  font-size: 20px;
+}
+// Add margin after input with prefix
+.mat-mdc-form-field-infix {
+  margin-left: 4px;
 }
